backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
google-generativeai==0.8.5
firebase-admin==6.4.0
python-dotenv==1.0.1
pydantic==2.6.1
Pillow==10.2.0
moviepy==1.0.3
torch
diffusers
transformers
accelerate
pypdf
python-docx
beautifulsoup4
youtube-transcript-api
requests
